Refactors string extension and modifies API logic
Changes the multiplier parameter type in the API request splitter to double for improved precision.

Introduces a new string extension providing camelCase conversion with validation for edge cases.

Removes obsolete files related to string inference functionality.

Adds comprehensive tests for string extension behaviors, ensuring robustness.

Improves code maintainability and prepares for future enhancements.

+++ lib/src/core/api_request_splitter.dart
@@ -13,7 +13,7 @@ import 'package:gobabel_core/gobabel_core.dart';
 /// @return A list of maps, each representing a manageable group for API requests
 List<Map<String, String>> splitIntoManageableGroupsForApi(
   Map<String, String> inputMap, [
-  int? multiplier,
+  double? multiplier,
 ]) {
   final refMultiplier = multiplier ?? kRequestMaxKeyMultiplaier;
   if (inputMap.isEmpty) return [];
